requirements.txt
asyncio
discord
python-dotenv
PyNaCl
supabase